Add sunxi armsoc glue
The Allwinner / sunxi platform does not have any special requirements
for GEMs, thus we can use DRM_IOCTL_MODE_CREATE_DUMB. The hardware does
not support true color hardware cursors, so we use a DRM plane for the
cursor instead. This plane is initialized with the highest zpos when
zpos is supported by the driver.

Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>
[wens@csie.org: Use DRM_IOCTL_MODE_CREATE_DUMB to create GEM]
[wens@csie.org: set zpos for cursor plane]
[wens@csie.org: write commit message]
Signed-off-by: Chen-Yu Tsai <wens@csie.org>