Add support for platform bus probing
Newer Xserver releases built with systemd support let systemd-logind
acquire the drm fd with drm master, which is then passed to the Xserver.

Unfortunately the old probe method does not provide a way for the server
to pass the fd through, nor does the server drop the drm master. This
results in the probe function unable to acquire the master, and failing
the probe.

The new "platform bus" probe path provides an option for the driver to
retrieve the master fd, and if the driver doesn't support this, the
server will drop its master so the driver can acquire it.

This patch only adds the platform probe function, without supporting
retrieving the drm master fd from the server.

Signed-off-by: Chen-Yu Tsai <wens@csie.org>