Select plane supporting ARGB8888 for cursor
The cursor uses ARGB8888, and if the driver is using a plane instead of
the cursor API for it, the plane needs to support ARGB8888 as well.

There is no guarantee that the first plane in the list supports it.
This is the case with the DRM driver for Allwinner's DE 2.0 mixers.
The first plane is the VI (video interface) layer, which supports
YUV formats, but not formats with alpha channel.

This patch makes the driver go through the list of planes and select
one that does support ARGB8888.

Signed-off-by: Chen-Yu Tsai <wens@csie.org>